% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapAndRandomize.R
\name{Dble.bootstrap}
\alias{Dble.bootstrap}
\title{Double bootstrap function}
\usage{
Dble.bootstrap(catch, varnames = c("Haul", "nwide", "nfine"),
  smooth = T)
}
\arguments{
\item{catch}{Stacked matrix or dataframe of catches in SELECT format}

\item{varnames}{Length-3 vector containing name of haulID variable, and 2 catch freq vars}

\item{smooth}{Smooth at within-haul phase to avoid losing degrees of freedom (from increasing number of zero freqs)}
}
\value{
Dataframe of double-bootstrapped freqs
}
\description{
Double bootstrap function
}
